agenttrinity/requirements.txt
Flask>=2.2
psycopg2-binary>=2.9
requests>=2.28
gunicorn>=20.1
python-dotenv>=1.0

